requirements.txt
fastapi
uvicorn
openai
python-dotenv
azure-search-documents
azure-core
azure-identity
pydantic
